+++ devito/ir/__init__.py
@@ -1,2 +1,3 @@
 from devito.ir.dfg import *  # noqa
 from devito.ir.clusters import *  # noqa
+from devito.ir.iet import *  # noqa
