+++ devito/ir/iet/__init__.py
@@ -1,0 +1,4 @@
+from devito.ir.iet.nodes import *  # noqa
+from devito.ir.iet.visitors import *  # noqa
+from devito.ir.iet.properties import *  # noqa
+from devito.ir.iet.utils import *  # noqa
