+++ devito/operator.py
@@ -7,26 +7,26 @@ import ctypes
 import numpy as np
 import sympy
 
-from devito.flow import analyze_iterations
+from devito.arguments import infer_dimension_values_tuple
 from devito.cgen_utils import Allocator
 from devito.compiler import jit_compile, load
 from devito.dimension import Dimension
 from devito.dle import compose_nodes, filter_iterations, transform
 from devito.dse import rewrite
+from devito.flow import analyze_iterations
+from devito.exceptions import InvalidArgument, InvalidOperator
 from devito.function import Forward, Backward, CompositeFunction
-from devito.ir import clusterize
-from devito.types import Object
 from devito.logger import bar, error, info
-from devito.nodes import Element, Expression, Callable, Iteration, List, LocalExpression
+from devito.ir.clusters import clusterize
+from devito.ir.iet import (Element, Expression, Callable, Iteration, List,
+                           LocalExpression, FindScopes, ResolveTimeStepping,
+                           SubstituteExpression, Transformer, NestedTransformer)
 from devito.parameters import configuration
 from devito.profiling import create_profile
 from devito.stencil import Stencil
 from devito.symbolics import indexify, retrieve_terminals
 from devito.tools import as_tuple, filter_sorted, flatten, numpy_to_ctypes, partial_order
-from devito.visitors import (FindScopes, ResolveTimeStepping,
-                             SubstituteExpression, Transformer, NestedTransformer)
-from devito.exceptions import InvalidArgument, InvalidOperator
-from devito.arguments import infer_dimension_values_tuple
+from devito.types import Object
 
 
 class Operator(Callable):
