+++ devito/profiling.py
@@ -7,9 +7,9 @@ from functools import reduce
 from ctypes import Structure, byref, c_double
 from cgen import Struct, Value
 
-from devito.nodes import Expression, TimedList
+from devito.ir.iet import (Expression, TimedList, IsPerfectIteration,
+                           FindSections, FindNodes, Transformer)
 from devito.symbolics import estimate_cost, estimate_memory
-from devito.visitors import IsPerfectIteration, FindSections, FindNodes, Transformer
 
 __all__ = ['Profile', 'create_profile']
