+++ tests/conftest.py
@@ -7,7 +7,7 @@ from sympy import cos, Symbol  # noqa
 from devito import (Dimension, Eq, TimeDimension, SteppingDimension, SpaceDimension,
                     Constant, Function, configuration)  # noqa
 from devito.types import Scalar, Array
-from devito.nodes import Iteration
+from devito.ir.iet import Iteration
 from devito.tools import as_tuple
