+++ tests/test_dle.py
@@ -13,9 +13,9 @@ from devito.dle import retrieve_iteration_tree, transform
 from devito.dle.backends import DevitoRewriter as Rewriter
 from devito import Grid, Function, TimeFunction, Eq, Operator
 from devito.flow import analyze_iterations
-from devito.nodes import ELEMENTAL, Expression, Callable, Iteration, List, tagger
-from devito.visitors import (ResolveTimeStepping, SubstituteExpression,
-                             Transformer, FindNodes)
+from devito.ir.iet import (ELEMENTAL, Expression, Callable, Iteration, List, tagger,
+                           ResolveTimeStepping, SubstituteExpression,
+                           Transformer, FindNodes)
 
 
 @pytest.fixture(scope="module")
